+++ Clients/HttpClient.cs
@@ -2,7 +2,6 @@
 // Wifft licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
 using System.Net.Http;
@@ -10,6 +9,8 @@ using System.Net.Http.Headers;
 using System.Text;
 using System.Text.Json;
 using System.Threading.Tasks;
+
+using Microsoft.Extensions.Logging;
 
 using WifftOCR.DataModels;
 using WifftOCR.Services;