+++ ViewModels/WelcomePageViewModel.cs
@@ -1,4 +1,8 @@
-﻿namespace WifftOCR.ViewModels
+﻿// Copyright (c) Wifft 2023
+// Wifft licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace WifftOCR.ViewModels
 {
     internal partial class WelcomePageViewModel
     {