+++ Views/WelcomePage.xaml.cs
@@ -1,13 +1,15 @@
 // Copyright (c) Wifft 2023
 // Wifft licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
+
 using System;
 
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
+using Microsoft.UI.Dispatching;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
-using Microsoft.UI.Dispatching;
+
 
 using WifftOCR.Providers;